Fix session/result lifetime issues.
We allow session to be closed/deleted while results from that session are still usable. The logic was not fully correct in such a situation. This fix adds more bookkeeping logic to make sure that "orphaned" result does not try to delete or access its session object (which might no longer exist).