CDK: Fix public header dependencies and other refactorings.
Make sure that the "general" CDK headers do not use types defined in more "specific" headers like cdk/mysqlx/*.h or cdk/protocol/*.h. The biggest change is to redefine cdk::Doc_path interface which, before the change, was taken from the protocol headers. Now it is a separate interface following the visitor pattern where document path components are reported via processor callbacks.

The code was updated to the new cdk::Doc_path interface, including the parser. On that occasion a couple of tweaks to the document path parsing logic was made. Also, the Token_op_base class was made a common base of the Expr_parser<> template (as these methods are used everywhere).