cdk: cmake infrastructure for building external cmake projects
For exteranl dependencies, such as protobuf and compression libraries, their bundled sources are built as a separate, external cmake project and import targets are defined to refer to the produced libraries in the main project. New cmake commands add_ext() and add_ext_targets() are defined to manage this setup. Also, a mechanism to cache intial cmake settings and this way speed-up configuration of external projects is implemented.